% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMg.R
\name{EMg}
\alias{EMg}
\title{EM solution of binary mixed Gaussian model.}
\usage{
EMg(e1, X, t_max = 100)
}
\arguments{
\item{e1}{a list consisting of the iterative initial value.}

\item{X}{a data matrix or data.frame.}

\item{t_max}{maximum number of iterations.}
}
\value{
a list consisting of mu ,sigma matrix and the The mixing ratio.
}
\description{
EM solution of binary mixed Gaussian model.
}
\examples{
\dontrun{
e1<-list()
e1$mu<-list(c(5,62),c(6,85))
e1$lambda<-c(.3,.7)
e1$sigma<-list(diag(2),diag(2))
e_hat<-EMg(e1,as.matrix(faithful),100)
}
}
