% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMLE.R
\name{pmle}
\alias{pmle}
\title{PMLE-penalized maximum likelihood estimate in linear regression model.}
\usage{
pmle(x, y, lam0 = NULL, LSE = FALSE)
}
\arguments{
\item{x}{predictors, an n by p matrix with n > 1 and p > 1}

\item{y}{response, an n-vector with n > 1}

\item{lam0}{Initial penalty level. Default is sqrt(2 * log(p)/n)}

\item{LSE}{If TRUE, compute least squares estimates after scaled Lasso selection. Default is FALSE}
}
\value{
\item{sigma}{the estimated noise level}
\item{coef}{the estimated coefficients}
\item{fitted.value}{the fitted value}
\item{residuals}{the residuals}
\item{lse}{the least square estimation after the seletion}
}
\description{
A penalized maximum likelihood estimator using R
}
\examples{
\dontrun{
data(X)
attach(X)
data(Y)
attach(Y)
x = X
y = Y
pmle(x,y,lam0=3)$coef
}
}
