% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIG.R
\name{RIG}
\alias{RIG}
\title{RIG for random number generation from Inverse Gaussian Distribution.}
\usage{
RIG(n, mu, lambda)
}
\arguments{
\item{n}{number of observations. If 'length(n) > 1', the length is taken to be the number required.}

\item{mu}{mean, it can be vector, each element must be bigger than 0}

\item{lambda}{shape parameter, it can be a vector each element must be bigger than 0}
}
\value{
Random number from Inverse Gaussian Distribution.
}
\description{
Generate random number from Inverse Gaussian Distribution for Bayesian linear regression.
}
\examples{
\dontrun{
n <- 100
mu <- c(1,1)
lambda <- c(2,4)
RIG(n,mu,lambda)
}
}
