\name{StatComp21010-package}
\alias{StatComp21010-package}
\alias{StatComp21010}
\docType{package}
\title{
  Three functions with homework answers
}
\description{
  This package includes three main functions RIG , PMLE and EMg .The function RIG is random number generation from Inverse Gaussian Distribution.The function PMLE is the L1 penalized maximum likelihood estimator(Stadler et al.,2010). And the function EMg is used to solve the EM estimation of binary mixed Gaussian model.
}
\author{
Shen Ao

Maintainer: Shen Ao <sa1014@mail.ustc.edu.cn>
}
\keyword{ package }
